MSBuild sample for example project
Add a MSBuild sample to the example project. To get this to work I
had to update the FluentMigrator dll and update the reference in
the FluentMigrator.Example project but I am not including that in
this pull request. Maybe something that should be added to the main
build process?