Throw exception if invalid processor parameter
If an incorrect parameter has been added for the provider then an
exception is thrown in TaskExecutor with a similar friendly error
message to the console one.

Added test for this and fixed the other TaskExecutor tests. Made
Initialize method virtual in TaskExecutor so that it could be
overriden for testing.