Defer Insert/Delete data value materialization until generation.
Previously, InsertDataDefinition and DeleteDataDefinition recorded
explicit values which could not be (easily) parsed or modified at the
generation stage.
This commit now employs ReflectedDataDefinition to hold a reference to
the anonymously typed object containing the data to insert/delete. This
ReflectedDataDefinition is then evaluated into a collection of
IDataValue (replacing the KeyValuePair<string,object>) instances by an
IEvaluator. Each generator now takes an IEvaluator as a constructor
parameter (in a similar way to the IColumn and IQuoter parameters) and a
GenericEvaluator is provided for all Generator types.
Finally a derived MySqlEvaluator is provided which allows the
MySqlGenerator to construct a SQL statement capable of inserting binary
data (via hex encoding and use if the MySql UNHEX function).