Added .editorconfig for sharing formatting
EditorConfig (see http://editorconfig.org/) is a way of sharing
Visual Studio settings for a solution. If you have the plugin installed
then it will enforce the rules in the .editorconfig file. We need this
so that new pull requests will be formatted with spaces instead of
tabs. It doesn't work if the committer does not install the plugin and
isn't meant as a requirement for submitting pull requests. It is more
intended to help those switching between projects with different tab
settings (for example me).

This will also work for Sublime Text, Emacs, Vim and Notepad++.