Add Command & Multiline Delimiters
When using FluentMigrator to create a script of migrations (for example
to provide to DBA's for deployment) MySQL requires each command to be
delimited with a semi-colon.

This check-in updates the SQL templates for all generators to provide an
extra delimiter placeholder. CommandDelimiter and MultilineDelimiters
are also provided as virtual getters and are substituted into the SQL
templates as appropriate.

This change required only limited modifications outside of
GenericGenerator and all unit tests continue to pass.