Fix issue with Alter Column for Oracle
Alter Column defaulted text to "Not Null" if the column was not nullable
(the default)
Alter.Column("SystemName")
.OnTable(AlertDefs)
.AsString(255);
would result in
ALTER TABLE stAlertDefs MODIFY SystemName NVARCHAR2(255) NOT NULL

This is not allowed by Oracle because the column is already not null.

Solution was to make IsNullable a Nullable<bool> and only add "Not Null"
to the alter if .NotNullable() was specified.  This should only affect
Oracle.