Makes tests for Sql Server sequences rollback on fail
If not Sql Server 2012 is enabled in the IntegrationTestOptions then
ignore these tests.

If the one of the tests fails then rollback the transaction. Otherwise
a schema can be created that messes up the other tests. The basic
principle is that a test should be repeatable and should cleanup after
itself.