use `ToList()`; remove short-circuit conditional
It's clear that `BeforeEach` and `AfterEach`  should not run, but it is
not clear that other, wrapping maintenance should not since the Profiles
could require them.  This may require more testing.  Would like to see
more feedback on the need for ApplyProfiles to run, absent any
underlying migrations.  Perhaps should push this back from 1.6.3?