Support renaming columns in MariaDb 10.2
MariaDb 10.2 returns the string 'NULL' for column default, instead of actual NULL as in previous versions. This was an intentional change by them, to remove some ambiguity between NULL as in "there is no column default" and NULL as in "NULL is the column default". https://jira.mariadb.org/browse/MDEV-13132

To support this change, the column definition when renaming a column should not quote NULL-strings from the column_default, as they actually mean that the default is NULL (not 'NULL').