Remove obsolete ExplicitUnicodeString
This was only supported for SQL Server, and the correct solution is to set-up the right data type.  For Update expressions, unicode will be assumed by default.  If your update expression uses a Where clause where this would cause an implicit conversion, use Execute.Sql instead, or Execute.WithConnection and use your data access library of choice to do the update.